requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
Werkzeug==3.0.1
PyJWT==2.8.0
requests-oauthlib==1.3.1
python-socketio==5.10.0
stripe==7.0.0
pillow==10.0.0
bcrypt==4.0.1
gevent==23.9.1
gevent-websocket==0.10.1
pytest==7.4.3
pytest-flask==1.3.0
